pyproject.toml
[tool.poetry]
name = "report_builder"
version = "0.1.0"
description = ""
authors = [" "]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.9.2"
"ruamel.yaml" = "0.17.40"
pyspark = "3.4.3"
requests = "^2.32.3"
confluent-kafka = "^2.6.0"
jsonschema = "^4.23.0"
pendulum = "^3.0.0"
yandex_tracker_client = "2.8"
pandas = "1.5.3"
urllib3 = "^1.0.0"
openpyxl = "^3.1.5"
numpy = "1.26.4"
XlsxWriter = "^3.2.0"
docxtpl = "^0.18.0"
botocore = "^1.35.65"

[tool.poetry.dev-dependencies]
isort = "^5.11.4"
yapf = "^0.32.0"
tomli = "^2.0.1"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"